+++ src/main/java/net/frozenblock/wilderwild/mixin/client/wind/FallingDustParticleMixin.java
@@ -47,5 +47,4 @@ public abstract class FallingDustParticleMixin extends SingleQuadParticle {
 		this.yd += wind.y * 0.00005D;
 		this.zd += wind.z * 0.001D;
 	}
-
 }