+++ src/main/java/net/frozenblock/wilderwild/mixin/client/wind/FallingLeavesParticleMixin.java
@@ -197,5 +197,4 @@ public abstract class FallingLeavesParticleMixin extends SingleQuadParticle {
 		}
 		return original;
 	}
-
 }