+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
-- Fixed a crash related to Wardens.
+- The position of clouds based on Wind is now extracted to the Render State before being used.